Revert "tweaks to move up to clang 8. higher versions of clang seem to not work even with API updates in coriander"
This reverts commit ea5ae493a8bd2ee6f0ed9e252818fd373d3095ea.

+++ src/mutations.cpp
@@ -49,7 +49,7 @@ void appendGlobalConstructorCall(Module *M, std::string functionName) {
         M->getOrInsertFunction(
             functionName,
             Type::getVoidTy(M->getContext()),
-            static_cast<size_t>(NULL)),
+            NULL),
         ConstantPointerNull::get(PointerType::get(IntegerType::get(M->getContext(), 8), 0))
     };
     initializers[oldNumConstructors] = ConstantStruct::getAnon(ArrayRef<Constant *>(&structValues[0], &structValues[3]));
